% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vol_bl76.R
\docType{data}
\name{df1_sample}
\alias{df1_sample}
\title{implied vol via Black76}
\format{An object of class \code{data.frame} with 2 rows and 5 columns.}
\usage{
df1_sample
}
\arguments{
\item{df_raw, }{is a dataframe with the same formats with the sample df1.}
}
\description{
compute implied vol and plot
}
\details{
this function is used for calculating the implied vol for European options
using the balck76 model and plot the figure of implied vol under diverse strike
prices for call options vs put options.
}
\examples{
plotImpliedVol(df1_sample)
plotImpliedVol()
}
\keyword{datasets}
